Updated file system
I reverted most of the code to a previous version to continue working on bugs. This current version includes 7 modes for laser and shutter control with live feed. The calibration mode should work also. I also fixed the timestamps and updated the file system so that each calibration creates a new folder labeled with the time and all following files are stored in the folder for the most recent calibration.